requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
alembic==1.12.1
redis==5.0.1
celery==5.3.4
pydantic>=2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pillow==10.4.0
google-generativeai==0.3.2
alibabacloud-ocr-api20210707==2.0.2
alibabacloud-tea-openapi>=0.3.7
oss2>=2.18.0
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1